+++ contracts/src/ModuleCRC20.sol
@@ -5,9 +5,17 @@ import "ds-token/token.sol";
 contract ModuleCRC20 is DSToken  {
     // sha256('cronos')[:20]
     address constant module_address = 0x6526B48f897f6e08067dF00A1821d783cbC2af8b;
+    string denom;
 
-    constructor(string memory denom, uint8 decimals_) DSToken(denom) public {
+    event __CronosSendToEthereum(address recipient, uint256 amount, uint256 bridge_fee);
+
+    constructor(string memory denom_, uint8 decimals_) DSToken(denom) public {
         decimals = decimals_;
+        denom = denom_;
+    }
+
+    function native_denom() public view returns (string memory) {
+        return denom;
     }
 
     function mint_by_cronos_module(address addr, uint amount) public {
@@ -22,5 +30,11 @@ contract ModuleCRC20 is DSToken  {
         balanceOf[addr] = sub(balanceOf[addr], amount);
         totalSupply = sub(totalSupply, amount);
         emit Burn(addr, amount);
+    }
+
+    // send to ethereum through gravity bridge
+    function send_to_ethereum(address recipient, uint amount, uint bridge_fee) public {
+        burn(msg.sender, add(amount, bridge_fee));
+        emit __CronosSendToEthereum(recipient, amount, bridge_fee);
     }
 }