+++ integration_tests/cosmoscli.py
@@ -272,9 +272,6 @@ class CosmosCLI:
                 "-y",
                 "--generate-only" if generate_only else None,
                 home=self.data_dir,
-                keyring_backend="test",
-                chain_id=self.chain_id,
-                node=self.node_rpc,
                 fees=fees,
             )
         )
@@ -943,7 +940,7 @@ class CosmosCLI:
             self.raw(
                 "tx",
                 "gravity",
-                "send-to-etheruem",
+                "send-to-ethereum",
                 receiver,
                 coins,
                 fee,