+++ integration_tests/test_basic.py
@@ -1,7 +1,38 @@
-from .utils import ADDRS
+from pathlib import Path
+
+from .utils import ADDRS, deploy_contract
 
 
 def test_basic(cluster):
     w3 = cluster.w3
     assert w3.eth.chain_id == 777
     assert w3.eth.get_balance(ADDRS["community"]) == 10000000000000000000000
+
+
+def test_native_call(cronos):
+    """
+    test contract native call on cronos network
+    - deploy test contract
+    - run native call, expect failure, becuase no native fund in contract
+    - send native tokens to contract account
+    - run again, expect success and check balance
+    """
+    w3 = cronos.w3
+    contract = deploy_contract(
+        w3,
+        Path(__file__).parent
+        / "contracts/artifacts/contracts/TestERC20A.sol/TestERC20A.json",
+    )
+
+    amount = 100
+
+    # the coinbase in web3 api is the same address as the validator account in
+    # cosmos api
+
+    # expect failure, because contract is not connected with native denom yet
+    # TODO complete the test after gov managed token mapping is implemented.
+    txhash = contract.functions.test_native_transfer(amount).transact(
+        {"from": w3.eth.coinbase}
+    )
+    receipt = w3.eth.wait_for_transaction_receipt(txhash)
+    assert receipt.status == 0, "should fail"